+++ mayan/apps/file_caching/urls.py
@@ -6,12 +6,11 @@ from .views import CacheListView, CachePurgeView
 
 urlpatterns = [
     url(
-        regex=r'^caches/$',
-        name='cache_list', view=CacheListView.as_view()
+        regex=r'^caches/$', name='cache_list', view=CacheListView.as_view()
     ),
     url(
-        regex=r'^caches/(?P<cache_id>\d+)/purge/$',
-        name='cache_purge', view=CachePurgeView.as_view()
+        regex=r'^caches/(?P<cache_id>\d+)/purge/$', name='cache_purge',
+        view=CachePurgeView.as_view()
     ),
     url(
         regex=r'^caches/multiple/purge/$', name='cache_multiple_purge',